python/requirement.txt
fastapi
joblib
pandas
torch
torchvision
uvicorn
pillow
python-multipart
scikit-learn
tensorflow
google-generativeai
python-dotenv
opencv-python
scipy
pydantic-settings
motor
pydantic[email]